requirements.txt
numpy==2.0.0
pandas==2.2.2
Pillow==10.3.0
PyQt5==5.15.10
PyQt5_sip==12.13.0
scikit_learn==1.4.2
